Refactor and standardise MenuGuideFragment
