+++ app/src/main/java/com/delacrixmorgan/kingscup/common/GameEngine.kt
@@ -17,7 +17,6 @@ import kotlin.collections.ArrayList
  */
 
 class GameEngine private constructor(context: Context) {
-
     companion object {
         const val GAME_ENGINE_TAUNT = "GAME_ENGINE_TAUNT"
         const val GAME_ENGINE_STATUS = "GAME_ENGINE_STATUS"
