--- app/src/main/res/layout/fragment_guide_list.xml
@@ -1,67 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@drawable/wallpaper_pink"
-    android:padding="50dp">
-
-    <RelativeLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_centerInParent="true">
-
-        <ImageView
-            android:id="@+id/symbolImageView"
-            android:layout_width="150dp"
-            android:layout_height="150dp"
-            android:layout_centerHorizontal="true"
-            android:layout_marginBottom="10dp"
-            android:layout_marginTop="10dp"
-            android:background="@drawable/heart_pink" />
-
-        <TextView
-            android:id="@+id/tv_guide_header"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_below="@id/symbolImageView"
-            android:gravity="center"
-            android:text="@string/quick_guide"
-            android:textAllCaps="true"
-            android:textColor="@android:color/white"
-            android:textSize="25sp"
-            android:textStyle="bold" />
-
-        <RelativeLayout
-            android:id="@+id/rl_guide_text"
-            android:layout_width="match_parent"
-            android:layout_height="250dp"
-            android:layout_below="@id/tv_guide_header">
-
-            <android.support.v4.view.ViewPager
-                android:id="@+id/viewPager"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_marginBottom="50dp"
-                android:layout_marginTop="20dp" />
-
-            <android.support.design.widget.TabLayout
-                android:id="@+id/tabLayout"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_alignParentBottom="true"
-                app:tabBackground="@drawable/tab_layout"
-                app:tabGravity="center"
-                app:tabIndicatorHeight="0dp" />
-        </RelativeLayout>
-
-        <Button
-            android:id="@+id/skipButton"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_below="@+id/rl_guide_text"
-            android:background="@drawable/btn_rounded_white"
-            android:text="@string/skip"
-            android:textColor="@color/colorPrimary" />
-    </RelativeLayout>
-</RelativeLayout>
