+++ app/src/main/res/layout/fragment_menu_setting.xml
@@ -56,6 +56,7 @@
                 android:orientation="vertical">
 
                 <RelativeLayout
+                    android:id="@+id/guideLayout"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
                     android:layout_marginEnd="24dp"
