+++ app/src/main/res/layout/fragment_menu_unlock.xml
@@ -34,7 +34,7 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:padding="10dp"
-            android:text="King's Cup Patron"
+            android:text="Pay What You Want"
             android:textAllCaps="true"
             android:textColor="@android:color/white"
             android:textSize="22sp" />
@@ -68,15 +68,6 @@
                 android:layout_height="wrap_content"
                 android:text="c) Remove Ads"
                 android:textAlignment="center" />
-
-            <TextView
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_marginTop="12dp"
-                android:text="Pay What You Want"
-                android:textAlignment="center"
-                android:textColor="@color/colorPrimary"
-                android:textSize="18sp" />
 
             <TextView
                 android:layout_width="match_parent"
