+++ app/src/main/res/layout/view_guide_list.xml
@@ -5,6 +5,6 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:gravity="center"
-    android:textColor="@android:color/white"
+    android:textColor="@color/colorPrimary"
     android:textSize="18sp"
     tool:text="@tools:sample/lorem" />
