+++ app/src/main/res/values/colors.xml
@@ -7,6 +7,7 @@
     <color name="colorPinkTheme">#ae275f</color>
     <color name="colorDivider">#EFEBE9</color>
 
+    <color name="pink">#FF4081</color>
     <color name="orange">#f47709</color>
     <color name="blue">#109cfb</color>
     <color name="green">#4CAF50</color>
